Added a container component for botNotOpenExplorer. Moved command service call to saga.
